✅ add test + refactor
