+++ src/handlers/entries/utils.rs
@@ -1,14 +1,21 @@
-use bigdecimal::BigDecimal;
+use bigdecimal::{BigDecimal, ToPrimitive};
 use chrono::NaiveDateTime;
 
 use crate::infra::repositories::entry_repository::MedianEntry;
 
 /// Converts a currency pair to a pair id.
+///
+/// e.g "btc" and "usd" to "BTC/USD"
 pub(crate) fn currency_pair_to_pair_id(quote: &str, base: &str) -> String {
     format!("{}/{}", quote.to_uppercase(), base.to_uppercase())
 }
 
 /// Computes the median price and time from a list of entries.
+/// The median price is computed as the median of the median prices of each entry.
+/// The median time is computed as the median of the times of each entry.
+/// The median is computed as the middle value of a sorted list of values.
+/// If the list has an even number of values, the median is computed as the average of the two middle values.
+/// If the list is empty, None is returned.
 pub(crate) fn compute_median_price_and_time(
     entries: &mut Vec<MedianEntry>,
 ) -> Option<(BigDecimal, NaiveDateTime)> {
@@ -34,4 +41,75 @@ pub(crate) fn compute_median_price_and_time(
     };
 
     Some((median_price, median_time))
+}
+
+/// Computes the volatility from a list of entries.
+/// The volatility is computed as the annualized standard deviation of the log returns.
+/// The log returns are computed as the natural logarithm of the ratio between two consecutive median prices.
+/// The annualized standard deviation is computed as the square root of the variance multiplied by 10^8.
+pub(crate) fn compute_volatility(entries: &Vec<MedianEntry>) -> f64 {
+    let mut values = Vec::new();
+    for i in 1..entries.len() {
+        if entries[i].median_price.to_f64().unwrap() > 0.0
+            && entries[i - 1].median_price.to_f64().unwrap() > 0.0
+            && (entries[i].time - entries[i - 1].time).num_seconds() > 0
+        {
+            let log_return = (entries[i].median_price.to_f64().unwrap()
+                / entries[i - 1].median_price.to_f64().unwrap())
+            .ln()
+            .powi(2);
+
+            let time = (entries[i].time - entries[i - 1].time)
+                .num_seconds()
+                .to_f64()
+                .unwrap()
+                / 31536000 as f64; // One year in seconds
+
+            values.push((log_return, time));
+        }
+    }
+
+    let variance: f64 = values.iter().map(|v| v.0 / v.1).sum::<f64>() / values.len() as f64;
+    let volatility = variance.sqrt() * 10_f64.powi(8);
+
+    volatility
+}
+
+#[test]
+fn test_volatility() {
+    let entries = vec![
+        MedianEntry {
+            time: chrono::NaiveDateTime::from_timestamp_opt(1640995200, 0).unwrap(),
+            median_price: bigdecimal::BigDecimal::from(47686),
+            source: "source".to_string(),
+        },
+        MedianEntry {
+            time: chrono::NaiveDateTime::from_timestamp_opt(1641081600, 0).unwrap(),
+            median_price: bigdecimal::BigDecimal::from(47345),
+            source: "source".to_string(),
+        },
+        MedianEntry {
+            time: chrono::NaiveDateTime::from_timestamp_opt(1641168000, 0).unwrap(),
+            median_price: bigdecimal::BigDecimal::from(46458),
+            source: "source".to_string(),
+        },
+        MedianEntry {
+            time: chrono::NaiveDateTime::from_timestamp_opt(1641254400, 0).unwrap(),
+            median_price: bigdecimal::BigDecimal::from(45897),
+            source: "source".to_string(),
+        },
+        MedianEntry {
+            time: chrono::NaiveDateTime::from_timestamp_opt(1641340800, 0).unwrap(),
+            median_price: bigdecimal::BigDecimal::from(43569),
+            source: "source".to_string(),
+        },
+    ];
+
+    // TODO: add more tests
+    // This value was computed using a python script
+    assert_eq!(
+        compute_volatility(&entries),
+        54594693.50567423,
+        "wrong volatility"
+    );
 }
